third_party/blink/web_tests/external/wpt/html/rendering/unmapped-attributes-expected.txt

This is a testharness.js-based test.
Found 8 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] <input hspace> should not be mapped to style marginLeft in CSS1Compat mode
  assert_equals: expected "0px" but got "200px"
[FAIL] <input hspace> should not be mapped to style marginLeft in BackCompat mode
  assert_equals: expected "0px" but got "200px"
[FAIL] <input hspace> should not be mapped to style marginRight in CSS1Compat mode
  assert_equals: expected "0px" but got "200px"
[FAIL] <input hspace> should not be mapped to style marginRight in BackCompat mode
  assert_equals: expected "0px" but got "200px"
[FAIL] <input vspace> should not be mapped to style marginTop in CSS1Compat mode
  assert_equals: expected "0px" but got "200px"
[FAIL] <input vspace> should not be mapped to style marginTop in BackCompat mode
  assert_equals: expected "0px" but got "200px"
[FAIL] <input vspace> should not be mapped to style marginBottom in CSS1Compat mode
  assert_equals: expected "0px" but got "200px"
[FAIL] <input vspace> should not be mapped to style marginBottom in BackCompat mode
  assert_equals: expected "0px" but got "200px"
Harness: the test ran to completion.
